# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "XXX\u5546\u57ceAPI",
        "version": "0.1"
    },
    "paths": {
        "/api/v1/auth": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "\u83b7\u53d6\u8bbf\u95eetoken",
                "description": "\u83b7\u53d6\u8bbf\u95eetoken",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "\u7528\u6237\u90ae\u7bb1",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "\u7528\u6237\u5bc6\u7801",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/v1/refresh": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "\u5237\u65b0\u8bbf\u95eetoken",
                "description": "\u5237\u65b0\u8bbf\u95eetoken",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "\u8bbf\u95eetoken",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/v1/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "\u83b7\u53d6\u6240\u6709\u7528\u6237\u4fe1\u606f",
                "description": "\u83b7\u53d6\u6240\u6709\u7528\u6237\u4fe1\u606f",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "\u6301\u6709\u7684token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "definitions": {},
    "tags": [
        {
            "name": "users",
            "description": "\u7528\u6237\u7ba1\u7406\u529f\u80fd"
        },
        {
            "name": "auth",
            "description": "\u7528\u6237\u8ba4\u8bc1\u529f\u80fd"
        }
    ]
}